fixed compiler errors

+++ cpp/new.cpp
@@ -21,7 +21,7 @@
  */
 #include <stdlib.h>
 
-void *operator new(unsigned size)
+void *operator new(size_t size)
 {
     return malloc(size);
 }
